+++ src/df-architect.ts
@@ -12,7 +12,7 @@ import { WallJoinSplit } from './walls/WallJoinSplit.js';
 import { AltLightOrigin } from './lights/AltLightOrigin.js';
 import { QuickColourPicker } from './lights/QuickColourPicker.js';
 import CaptureGameScreen from './general/CaptureGameScreen.js';
-import { LightTemplate, LightingLayerOverride } from './lights/LightTemplate.js';
+import { LightTemplateManager, LightingLayerOverride } from './lights/LightTemplate.js';
 
 Hooks.once('init', function () {
 	if (!game.modules.get('lib-wrapper')?.active) return;
@@ -23,7 +23,6 @@ Hooks.once('init', function () {
 	WallJoinSplit.init();
 	AltLightOrigin.init();
 	CaptureGameScreen.init();
-	LightTemplate.init();
 });
 
 Hooks.once('setup', function () {
@@ -44,5 +43,6 @@ Hooks.once('ready', function () {
 	WallCtrlInvert.ready();
 	AltLightOrigin.ready();
 	QuickColourPicker.ready();
+	LightTemplateManager.ready();
 	LightingLayerOverride.ready();
 });