+++ src/lights/LightTemplate.ts
@@ -5,32 +5,15 @@ declare global {
 	const LightTemplates: _LightTemplates;
 }
 
+
 class _LightTemplates {
-	private static readonly FLAG_IS_TEMPLATE = 'isLightTemplate';
-
-	private _currentActiveTemplate: string | null = null;
-	get currentTemplate(): Partial<AmbientLight.Data> | null {
-		if (this._currentActiveTemplate === null) return null;
-		return this._extractLightDataFromMacroCommand(game.macros.get(this._currentActiveTemplate).data.command);
-	}
-
-	init() {
-		Hooks.on('renderLightConfig', this._renderLightConfig.bind(this));
-		Hooks.on('renderMacroConfig', this._renderMacroConfig.bind(this));
-		Hooks.on('renderSceneControls', async controls => {
-			if (!game.user.isGM) return;
-			if (controls.activeControl !== 'lighting') {
-				await this.deactivate();
-				return;
-			}
-		});
-		libWrapper.register(ARCHITECT.MOD_NAME, 'Hotbar.prototype._contextMenu', this._contextMenu, 'OVERRIDE');
-	}
-
-	async activate(macroId: string, lightData: Partial<AmbientLight.Data>) {
-		console.debug(`Run ${macroId} using ${JSON.stringify(lightData)}`)
-
-		if (this._currentActiveTemplate !== null) {
+	/**
+	 * Activates the given Light Template provided by the Macro ID.
+	 * @param macroId ID of Macro containing the light template.
+	 */
+	async activate(macroId: string) {
+		if (!game.user.isGM) return;
+		if (LightTemplateManager.currentActiveTemplate !== null) {
 			this.deactivate();
 		}
 
@@ -39,16 +22,20 @@ class _LightTemplates {
 			$('li.scene-control[data-control="lighting"]').trigger('click');
 		}
 
-		const lightTypes = Object.entries(CONST.SOURCE_TYPES).reduce((obj: { [key: string]: string }, e) => {
-			obj[e[1]] = `LIGHT.Type${e[0].titleCase()}`;
-			return obj;
-		}, {});
+		LightTemplateManager.currentActiveTemplate = macroId;
+		const lightData = LightTemplateManager.getCurrentTemplateData();
+		// This is just to make the colour text more visible
+		const tintColour = parseInt('0x1' + (lightData.tintColor || '#000000').substr(1));
 		const templateHtml = $(await renderTemplate(`modules/${ARCHITECT.MOD_NAME}/templates/cur-light-template.hbs`, {
 			name: game.macros.get(macroId).name,
 			data: lightData,
 			colorIntensity: Math.sqrt(lightData.tintAlpha).toNearest(0.05),
-			lightType: lightTypes[lightData.t],
-			animationType: lightData.lightAnimation.type === "" ? 'None' : CONFIG.Canvas.lightAnimations[lightData.lightAnimation.type].label
+			lightType: LightTemplateManager.lightTypes[lightData.t],
+			animationType: lightData.lightAnimation.type === "" ? 'None' : CONFIG.Canvas.lightAnimations[lightData.lightAnimation.type].label,
+			tintColorValue: lightData.tintColor || 'transparent',
+			tintColorLabel: lightData.tintColor || '#------',
+			// Determines if the tint colour is dark or light, the HBS template will change the text colour accordingly
+			tintIsDark: ((((tintColour & 0xff0000) >> 16) + ((tintColour & 0xff00) >> 8) + (tintColour & 0xff)) / 3) < 128
 		}));
 		const rect = $('#controls .active .control-tools')[0].getBoundingClientRect();
 		templateHtml.css('left', rect.right + 10 + 'px');
@@ -56,30 +43,82 @@ class _LightTemplates {
 		const buttons = templateHtml.find('a');
 		buttons.first().on('click', () => game.macros.get(macroId).sheet.render(true));
 		buttons.last().on('click', this.deactivate.bind(this));
-
-		this._currentActiveTemplate = macroId;
 		templateHtml.appendTo(document.body);
 	}
 
-	async deactivate() {
+	deactivate() {
+		if (!game.user.isGM) return;
 		// clear the status from previous selection
 		$('#dfarch-cur-light-template').remove();
-		this._currentActiveTemplate = null;
+		LightTemplateManager.currentActiveTemplate = null;
+	}
+}
+
+// @ts-ignore
+window.LightTemplates = new _LightTemplates();
+
+export class LightTemplateManager {
+	private static readonly FLAG_IS_TEMPLATE = 'isLightTemplate';
+
+	static currentActiveTemplate: string | null = null;
+	static getCurrentTemplateData(): Partial<AmbientLight.Data> | null {
+		if (this.currentActiveTemplate === null) return null;
+		return this._extractLightDataFromMacroCommand(game.macros.get(this.currentActiveTemplate).data.command);
 	}
 
-	private _renderLightConfig(app: LightConfig, html: JQuery<HTMLElement>, data: any) {
+	static get lightTypes(): { [key: string]: string } {
+		return Object.entries(CONST.SOURCE_TYPES).reduce((obj: { [key: string]: string }, e) => {
+			obj[e[1]] = `LIGHT.Type${e[0].titleCase()}`;
+			return obj;
+		}, {});
+	}
+
+	static ready() {
+		if (!game.user.isGM) return;
+		Hooks.on('renderLightConfig', this._renderLightConfig.bind(this));
+		Hooks.on('renderMacroConfig', this._renderMacroConfig.bind(this));
+		Hooks.on('renderSceneControls', controls => {
+			if (game.user.isGM && controls.activeControl !== 'lighting')
+				LightTemplates.deactivate();
+		});
+		Hooks.on('renderMacroDirectory', (app: MacroDirectory, html: JQuery<HTMLElement>, data: any) => {
+			const createTemplateButton = $(`<button style="flex:unset"><i class="far fa-lightbulb"></i>
+${'DF_ARCHITECT.LightTemplate_CreateTemplateButton_MacroDirectory'.localize()}</button>`);
+			createTemplateButton.appendTo('div.header-actions.action-buttons');
+			createTemplateButton.on('click', async () => await this._createTemplate({
+				t: 'l',
+				dim: 30,
+				angle: 360,
+				bright: 15,
+				rotation: 0,
+				tintColor: "",
+				tintAlpha: Math.pow(0.7, 2).toNearest(0.01),
+				darknessThreshold: 0,
+				lightAnimation: {
+					intensity: 5,
+					speed: 5,
+				},
+			}));
+		});
+		libWrapper.register(ARCHITECT.MOD_NAME, 'Hotbar.prototype._contextMenu', this._contextMenu, 'OVERRIDE');
+	}
+
+	private static _renderLightConfig(app: LightConfig, html: JQuery<HTMLElement>, data: any) {
 		const button = $(`<button type="button" style="margin-bottom:0.25em" name="save-template">
 <i class="far fa-lightbulb"></i>
-${'DF_ARCHITECT.LightTemplate_CreateTemplate_Button'.localize()}
+${'DF_ARCHITECT.LightTemplate_CreateTemplateButton_LightConfig'.localize()}
 </button>`);
 		html.find('button[name="submit"]').before(button);
-		button.on('click', (event) => { event.preventDefault(); this._createTemplate(app) });
+		button.on('click', (event) => {
+			event.preventDefault();
+			this._createTemplate(duplicate((<AmbientLight>app.object).data) as Partial<AmbientLight.Data>);
+		});
 	}
 
-	private async _renderMacroConfig(app: MacroConfig, html: JQuery<HTMLElement>, data: BaseEntitySheet.Data) {
+	private static async _renderMacroConfig(app: MacroConfig, html: JQuery<HTMLElement>, data: BaseEntitySheet.Data) {
 		console.debug(data.entity);
 		const macro = new Macro(data.entity as Macro.Data);
-		if (!macro.getFlag(ARCHITECT.MOD_NAME, _LightTemplates.FLAG_IS_TEMPLATE))
+		if (!macro.getFlag(ARCHITECT.MOD_NAME, this.FLAG_IS_TEMPLATE))
 			return;
 		html.find('div.form-group').remove();
 		html.find('footer').remove();
@@ -89,23 +128,18 @@ ${'DF_ARCHITECT.LightTemplate_CreateTemplate_Button'.localize()}
 		for (let [k, v] of Object.entries(CONFIG.Canvas.lightAnimations)) {
 			animationTypes[k] = v.label;
 		}
-		const lightTypes = Object.entries(CONST.SOURCE_TYPES).reduce((obj: { [key: string]: string }, e) => {
-			obj[e[1]] = `LIGHT.Type${e[0].titleCase()}`;
-			return obj;
-		}, {});
 
 		const lightData = this._extractLightDataFromMacroCommand((data.entity as Macro.Data).command);
 		const htmlData = {
 			object: duplicate(lightData),
-			lightTypes: lightTypes,
+			lightTypes: this.lightTypes,
 			lightAnimations: animationTypes,
 			colorIntensity: Math.sqrt(lightData.tintAlpha).toNearest(0.05)
 		}
 
-		// Render the form
-		const content = $(await renderTemplate(`modules/${ARCHITECT.MOD_NAME}/templates/light-template.hbs`, htmlData));
-
-		html.find('header.sheet-header').after(content);
+		// Render the form html
+		html.find('header.sheet-header')
+			.after(await renderTemplate(`modules/${ARCHITECT.MOD_NAME}/templates/light-template.hbs`, htmlData));
 
 		// @ts-ignore
 		// Replace the MacroConfig's update function with our own
@@ -122,11 +156,9 @@ ${'DF_ARCHITECT.LightTemplate_CreateTemplate_Button'.localize()}
 				name: name,
 				img: img
 			});
-			// If we are editing the currently selected light template
-			if (this._currentActiveTemplate === macro.id) {
-				// Update the UI
-				await this.activate(macro.id, formData);
-			}
+			// If we are editing the currently selected light template, update the UI
+			if (this.currentActiveTemplate === macro.id)
+				await LightTemplates.activate(macro.id);
 		}
 
 		// Resizes the config menu to accommodate the new elements.
@@ -139,8 +171,7 @@ ${'DF_ARCHITECT.LightTemplate_CreateTemplate_Button'.localize()}
 	 * Simple function used to generate a specialized Macro to be used for Light Templating.
 	 * @param app LightConfig Application
 	 */
-	private async _createTemplate(app: LightConfig) {
-		const lightData = duplicate((<AmbientLight>app.object).data) as Partial<AmbientLight.Data>;
+	private static async _createTemplate(lightData: Partial<AmbientLight.Data>) {
 		// Delete the obsolete fields
 		delete lightData._id;
 		delete lightData.x;
@@ -155,46 +186,38 @@ ${'DF_ARCHITECT.LightTemplate_CreateTemplate_Button'.localize()}
 			type: 'script',
 			command: this._generateCommandData(lightData)
 		});
-		await macro.setFlag(ARCHITECT.MOD_NAME, _LightTemplates.FLAG_IS_TEMPLATE, true);
+		await macro.setFlag(ARCHITECT.MOD_NAME, this.FLAG_IS_TEMPLATE, true);
 		macro.sheet.render(true);
 	}
 
-	private _generateCommandData(lightData: Partial<AmbientLight.Data>) {
-		return `const ld=${JSON.stringify(lightData, null, '')};\nLightTemplates.activate(this.id,ld);`
-	}
-
-	private _contextMenu(html: JQuery<HTMLElement>) {
+	/**
+	 * Unfortunately, by the design of this core function, I must duplicate the
+	 * code here in order to add a custom option.
+	 */
+	private static _contextMenu(html: JQuery<HTMLElement>) {
 		new ContextMenu(html, ".macro", [
 			{
 				name: "MACRO.Edit",
 				icon: '<i class="fas fa-edit"></i>',
 				condition: li => {
 					const macro = game.macros.get(li.data("macro-id"));
-					return macro && !macro.getFlag(ARCHITECT.MOD_NAME, _LightTemplates.FLAG_IS_TEMPLATE) ? macro.owner : false;
+					return macro && !macro.getFlag(ARCHITECT.MOD_NAME, this.FLAG_IS_TEMPLATE) ? macro.owner : false;
 				},
-				callback: li => {
-					const macro = game.macros.get(li.data("macro-id"));
-					macro.sheet.render(true);
-				}
+				callback: li => game.macros.get(li.data("macro-id")).sheet.render(true)
 			},
 			{
 				name: "DF_ARCHITECT.LightTemplate_EditTemplate_Label",
 				icon: '<i class="fas fa-edit"></i>',
 				condition: li => {
 					const macro = game.macros.get(li.data("macro-id"));
-					return macro && macro.getFlag(ARCHITECT.MOD_NAME, _LightTemplates.FLAG_IS_TEMPLATE) ? macro.owner : false;
+					return macro && macro.getFlag(ARCHITECT.MOD_NAME, this.FLAG_IS_TEMPLATE) ? macro.owner : false;
 				},
-				callback: li => {
-					const macro = game.macros.get(li.data("macro-id"));
-					macro.sheet.render(true);
-				}
+				callback: li => game.macros.get(li.data("macro-id")).sheet.render(true)
 			},
 			{
 				name: "MACRO.Remove",
 				icon: '<i class="fas fa-times"></i>',
-				callback: li => {
-					game.user.assignHotbarMacro(null, li.data("slot"));
-				}
+				callback: li => game.user.assignHotbarMacro(null, li.data("slot"))
 			},
 			{
 				name: "MACRO.Delete",
@@ -215,12 +238,17 @@ ${'DF_ARCHITECT.LightTemplate_CreateTemplate_Button'.localize()}
 		]);
 	}
 
-	private _extractLightDataFromMacroCommand(commandString: string): Partial<AmbientLight.Data> {
+	private static _generateCommandData(lightData: Partial<AmbientLight.Data>) {
+		return `const ld=${JSON.stringify(lightData, null, '')};\nLightTemplates.activate(this.id,ld);`
+	}
+	private static _extractLightDataFromMacroCommand(commandString: string): Partial<AmbientLight.Data> {
 		return JSON.parse(/const ld=(.+);/.exec(commandString)[1]) as Partial<AmbientLight.Data>
 	}
 }
 
-
+/**
+ * Set of functions that override the Mouse Events for the LightingLayer in order to place templated lights.
+ */
 export class LightingLayerOverride {
 	static ready() {
 		libWrapper.register(ARCHITECT.MOD_NAME, 'LightingLayer.prototype._onDragLeftStart',
@@ -235,8 +263,8 @@ export class LightingLayerOverride {
 	}
 
 	static _onClickLeft(this: LightingLayer, wrapper: Function, event: PIXI.InteractionEvent) {
-		const templateData = LightTemplates.currentTemplate;
-		// If we do not have an active template, pass through to original function
+		const templateData = LightTemplateManager.getCurrentTemplateData();
+		// If we do not have an active template/Ctrl not pressed, pass through to original function
 		if (templateData === null || !event.data.originalEvent.ctrlKey)
 			return wrapper(event);
 		const origin: { x: number, y: number } = (<any>event.data).origin;
@@ -244,22 +272,20 @@ export class LightingLayerOverride {
 	}
 
 	static _onClickRight(this: LightingLayer, wrapper: Function, event: PIXI.InteractionEvent) {
-		const templateData = LightTemplates.currentTemplate;
-		// If we do not have an active template, pass through to original function
+		const templateData = LightTemplateManager.getCurrentTemplateData();
+		// If we do not have an active template/Ctrl not pressed, pass through to original function
 		if (templateData === null || !event.data.originalEvent.ctrlKey)
 			return wrapper(event);
 		LightTemplates.deactivate();
 	}
 
 	static _onDragLeftStart(this: LightingLayer, wrapper: Function, event: PIXI.InteractionEvent) {
-		const templateData = LightTemplates.currentTemplate;
+		const templateData = LightTemplateManager.getCurrentTemplateData();
 		// If we do not have an active template, pass through to original function
 		if (templateData === null)
 			return wrapper(event);
-		// @ts-ignore
-		// PlaceablesLayer.prototype._onDragLeftStart(event);
 		const origin: { x: number, y: number } = (<any>event.data).origin;
-		// Create the preview source
+		// Create the preview source from the Template Data
 		const preview = new AmbientLight(mergeObject(templateData, { x: origin.x, y: origin.y }));
 		(<any>event.data).preview = this.preview.addChild(preview);
 		this.sources.set(preview.sourceId, preview.source);
@@ -267,8 +293,3 @@ export class LightingLayerOverride {
 		return preview.draw();
 	}
 }
-
-// @ts-ignore
-window.LightTemplates = new _LightTemplates();
-
-export const LightTemplate = LightTemplates;